# docs/api/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/users/{id}/role":{"put":{"tags":["用户管理"],"summary":"修改用户角色","operationId":"updateRole","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleRequest"}}},"required":true},"responses":{"200":{"description":"成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RUserInfoResponse"}}}},"404":{"description":"用户不存在","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}},"403":{"description":"权限不足","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/products/{id}":{"get":{"tags":["商品管理"],"summary":"获取商品详情","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RProductResponse"}}}}}},"put":{"tags":["商品管理"],"summary":"更新商品","operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RProductResponse"}}}}}},"delete":{"tags":["商品管理"],"summary":"删除商品","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/dealers/{id}":{"get":{"tags":["经销商管理"],"summary":"获取经销商详情","operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RDealerResponse"}}}}}},"put":{"tags":["经销商管理"],"summary":"更新经销商","operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DealerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RDealerResponse"}}}}}},"delete":{"tags":["经销商管理"],"summary":"删除经销商","operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/categories/{id}":{"get":{"tags":["商品分类管理"],"summary":"获取分类详情","operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RCategoryResponse"}}}}}},"put":{"tags":["商品分类管理"],"summary":"更新分类","operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RCategoryResponse"}}}}}},"delete":{"tags":["商品分类管理"],"summary":"删除分类","operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/products":{"get":{"tags":["商品管理"],"summary":"商品列表（分页）","operationId":"list","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"keyword","in":"query","required":false,"schema":{"type":"string"}},{"name":"categoryId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"status","in":"query","required":false,"schema":{"type":"string","enum":["DRAFT","ENABLED","DISABLED"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultProductResponse"}}}}}},"post":{"tags":["商品管理"],"summary":"创建商品","operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RProductResponse"}}}}}}},"/api/dealers":{"get":{"tags":["经销商管理"],"summary":"经销商列表（分页）","operationId":"list_1","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultDealerResponse"}}}}}},"post":{"tags":["经销商管理"],"summary":"创建经销商（关联用户）","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DealerRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RDealerResponse"}}}}}}},"/api/dealers/{id}/prices":{"get":{"tags":["经销商管理"],"summary":"查询经销商价格","operationId":"getPrices","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"productId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RListMapStringObject"}}}}}},"post":{"tags":["经销商管理"],"summary":"设置经销商专属价格","operationId":"setPrices","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BatchPriceRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/categories":{"get":{"tags":["商品分类管理"],"summary":"分类列表（全部）","operationId":"list_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RListCategoryResponse"}}}}}},"post":{"tags":["商品分类管理"],"summary":"创建分类","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RCategoryResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["认证管理"],"summary":"用户注册","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"注册成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RLoginResponse"}}}},"409":{"description":"用户名已存在","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}},"400":{"description":"参数校验失败","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/auth/refresh":{"post":{"tags":["认证管理"],"summary":"刷新 Token","operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"刷新成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RLoginResponse"}}}},"401":{"description":"refreshToken 无效或已过期","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/auth/login":{"post":{"tags":["认证管理"],"summary":"用户登录","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"登录成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RLoginResponse"}}}},"401":{"description":"用户名或密码错误","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/products/{id}/stock":{"patch":{"tags":["商品管理"],"summary":"更新库存","operationId":"updateStock","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"skuId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"stock","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RProductResponse"}}}}}}},"/api/products/{id}/enable":{"patch":{"tags":["商品管理"],"summary":"上架商品","operationId":"enable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RProductResponse"}}}}}}},"/api/products/{id}/disable":{"patch":{"tags":["商品管理"],"summary":"下架商品","operationId":"disable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RProductResponse"}}}}}}},"/api/dealers/{id}/toggle":{"patch":{"tags":["经销商管理"],"summary":"启用/禁用经销商","operationId":"toggle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RDealerResponse"}}}}}}},"/api/categories/{id}/toggle":{"patch":{"tags":["商品分类管理"],"summary":"启用/禁用分类","operationId":"toggle_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RCategoryResponse"}}}}}}},"/api/users":{"get":{"tags":["用户管理"],"summary":"用户列表（分页）","operationId":"list_3","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageResultUserInfoResponse"}}}},"403":{"description":"权限不足","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/users/{id}":{"get":{"tags":["用户管理"],"summary":"获取单个用户","operationId":"getById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RUserInfoResponse"}}}},"404":{"description":"用户不存在","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}},"403":{"description":"权限不足","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}},"delete":{"tags":["用户管理"],"summary":"删除用户","operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}},"404":{"description":"用户不存在","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}},"403":{"description":"权限不足","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}},"/api/health":{"get":{"tags":["health-controller"],"summary":"健康检查","operationId":"health","responses":{"200":{"description":"服务正常","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RMapStringObject"}}}}}}},"/api/auth/me":{"get":{"tags":["认证管理"],"summary":"获取当前用户信息","operationId":"me","responses":{"200":{"description":"成功","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RUserInfoResponse"}}}},"401":{"description":"未登录或Token无效","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}},"404":{"description":"用户不存在","content":{"*/*":{"schema":{"$ref":"#/components/schemas/R"}}}}}}}},"components":{"schemas":{"R":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object"}},"required":["code","message"]},"UpdateRoleRequest":{"type":"object","properties":{"role":{"type":"string","enum":["ADMIN","DEALER","CUSTOMER"]}},"required":["role"]},"RUserInfoResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/UserInfoResponse"}},"required":["code","message"]},"UserInfoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"type":"string"},"nickname":{"type":"string"},"avatar":{"type":"string"},"role":{"type":"string"}},"required":["nickname","role","username"]},"ProductRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"description":{"type":"string"},"categoryId":{"type":"string","format":"uuid"},"imageUrl":{"type":"string","maxLength":200,"minLength":0},"images":{"type":"array","items":{"type":"string"}},"hasSpecs":{"type":"boolean"},"skus":{"type":"array","items":{"$ref":"#/components/schemas/SkuItem"}}},"required":["categoryId","hasSpecs","imageUrl","images","name","skus"]},"SkuItem":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"specs":{"type":"string"},"price":{"type":"number"},"stock":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"},"sortOrder":{"type":"integer","format":"int32"},"id":{"type":"string","format":"uuid"}},"required":["name","price","sortOrder","specs","stock"]},"ProductResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"categoryId":{"type":"string","format":"uuid"},"categoryName":{"type":"string"},"imageUrl":{"type":"string"},"images":{"type":"array","items":{"type":"string"}},"hasSpecs":{"type":"boolean"},"status":{"type":"string","enum":["DRAFT","ENABLED","DISABLED"]},"minPrice":{"type":"number"},"totalStock":{"type":"integer","format":"int32"},"skus":{"type":"array","items":{"$ref":"#/components/schemas/SkuResponse"}},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["categoryId","categoryName","hasSpecs","id","imageUrl","images","name","skus","status","totalStock"]},"RProductResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ProductResponse"}},"required":["code","message"]},"SkuResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"specs":{"type":"string"},"price":{"type":"number"},"stock":{"type":"integer","format":"int32"},"imageUrl":{"type":"string"},"sortOrder":{"type":"integer","format":"int32"}},"required":["id","name","price","sortOrder","specs","stock"]},"DealerRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"minLength":0},"userId":{"type":"string","format":"uuid"},"phone":{"type":"string","maxLength":20,"minLength":0},"address":{"type":"string","maxLength":200,"minLength":0},"contactPerson":{"type":"string","maxLength":100,"minLength":0}},"required":["name","userId"]},"DealerResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"username":{"type":"string"},"name":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"contactPerson":{"type":"string"},"enabled":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["enabled","id","name","userId","username"]},"RDealerResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/DealerResponse"}},"required":["code","message"]},"CategoryRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":50,"minLength":0},"description":{"type":"string","maxLength":200,"minLength":0},"sortOrder":{"type":"integer","format":"int32"}},"required":["name","sortOrder"]},"CategoryResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"description":{"type":"string"},"sortOrder":{"type":"integer","format":"int32"},"enabled":{"type":"boolean"},"productCount":{"type":"integer","format":"int64"}},"required":["enabled","id","name","productCount","sortOrder"]},"RCategoryResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/CategoryResponse"}},"required":["code","message"]},"BatchPriceRequest":{"type":"object","properties":{"prices":{"type":"array","items":{"$ref":"#/components/schemas/DealerPriceRequest"}}},"required":["prices"]},"DealerPriceRequest":{"type":"object","properties":{"productId":{"type":"string","format":"uuid"},"skuId":{"type":"string","format":"uuid"},"price":{"type":"number"}},"required":["price","productId"]},"RegisterRequest":{"type":"object","properties":{"username":{"type":"string","maxLength":50,"minLength":3},"password":{"type":"string","maxLength":100,"minLength":6},"nickname":{"type":"string"}},"required":["password","username"]},"LoginResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"user":{"$ref":"#/components/schemas/UserInfoResponse"}},"required":["accessToken","user"]},"RLoginResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/LoginResponse"}},"required":["code","message"]},"RefreshTokenRequest":{"type":"object","properties":{"refreshToken":{"type":"string","minLength":1}},"required":["refreshToken"]},"LoginRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["password","username"]},"PageResultUserInfoResponse":{"type":"object","properties":{"list":{"type":"array","items":{"$ref":"#/components/schemas/UserInfoResponse"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}},"required":["list","page","size","total"]},"PageResultProductResponse":{"type":"object","properties":{"list":{"type":"array","items":{"$ref":"#/components/schemas/ProductResponse"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}},"required":["list","page","size","total"]},"RMapStringObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"object","additionalProperties":{"type":"object"}}},"required":["code","message"]},"PageResultDealerResponse":{"type":"object","properties":{"list":{"type":"array","items":{"$ref":"#/components/schemas/DealerResponse"}},"total":{"type":"integer","format":"int64"},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"}},"required":["list","page","size","total"]},"RListMapStringObject":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"type":"object","additionalProperties":{"type":"object"}}}},"required":["code","message"]},"RListCategoryResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponse"}}},"required":["code","message"]}}}}